+++ WorkflowProject/trunc/src/java/ru/sgnhp/web/UploadController.java
@@ -46,7 +46,7 @@ public class UploadController extends SimpleFormController {
             wf.setParentUserUid(initiator.getUid());
             wf.setUserUid(Long.valueOf(uid));
             wf.setDescription(task.getDescription());
-            wf.setState(Short.valueOf("0"));
+            wf.setState("0");
             wf.setAssignDate(task.getStartDate());
             workflowManagerService.assignTaskToUser(wf);
         }
